Changed the Tgam and Tcyt state change processing to ignore any macrophage that has already died on the same time step. Otherwise the macrophage would be killed twice making the statistics less accurate.
git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@712 88ce85bb-8564-4ae6-884f-ab7503230054

+++ simulation/tgamma.cpp
@@ -94,23 +94,32 @@ void Tgam::handleActive(const int, GrGrid& grid, GrStat& stats)
 {
 	GridCell& cell = grid(_row, _col);
 	
-	// get the macrophage
-	Mac* pMac = dynamic_cast<Mac*>(cell.getAgent(0));
-	if (!pMac) pMac = dynamic_cast<Mac*>(cell.getAgent(1));
-
-	_nextState = TGAM_ACTIVE;
-
-	// Fas/FasL induced apoptosis with probability 
-	if (pMac &&
-		(pMac->getState() == MAC_INFECTED || pMac->getState() == MAC_CINFECTED) &&
-		g_Rand.getReal() < _PARAM(PARAM_TGAM_PROB_APOPTOSIS_FAS_FASL))
+	if (cell.hasMac())
 	{
-		stats.incApoptosisFasFasL();
-		pMac->apoptosis(grid);
-		pMac->kill();
+		// get the macrophage
+		Mac* pMac = dynamic_cast<Mac*>(cell.getAgent(0));
+		if (!pMac) pMac = dynamic_cast<Mac*>(cell.getAgent(1));
 
-		/* Does Fas/FasL induced apoptosis contribute to caseation? */
-		cell.incNrKillings();
+		// If the mac died on this time step ignore it.
+		if (pMac->getNextState() == MAC_DEAD)
+		{
+			return;
+		}
+
+		_nextState = TGAM_ACTIVE;
+
+		// Fas/FasL induced apoptosis with probability
+		if (pMac &&
+			(pMac->getState() == MAC_INFECTED || pMac->getState() == MAC_CINFECTED) &&
+			g_Rand.getReal() < _PARAM(PARAM_TGAM_PROB_APOPTOSIS_FAS_FASL))
+		{
+			stats.incApoptosisFasFasL();
+			pMac->apoptosis(grid);
+			pMac->kill();
+
+			/* Does Fas/FasL induced apoptosis contribute to caseation? */
+			cell.incNrKillings();
+		}
 	}
 }
